+++ sema4ai/codegen/commands.py
@@ -1019,18 +1019,6 @@ COMMANDS = [
         server_handled=False,
         hide_from_command_palette=True,
     ),
-    Command(
-        "sema4ai.startDataServer",
-        "Start Data Server",
-        server_handled=False,
-        add_to_package_json=True,
-    ),
-    Command(
-        "sema4ai.stopDataServer",
-        "Stop Data Server",
-        server_handled=False,
-        add_to_package_json=True,
-    ),
 ]